social card and event cards changes
