+++ app/components/common/eventCard/index.test.tsx
@@ -1,108 +1,64 @@
-import { render } from '@testing-library/react';
-import dayjs from 'dayjs';
-import EventCard from './index';
+import { render, screen } from '@testing-library/react';
+import EventCard from '.';
 
-describe('EventCard', () => {
-  const date = dayjs().format('D');
-  const month = dayjs().format('MMM');
-  const dummyevent1 = {
-    id: 1,
-    title: 'timed event',
-    start: dayjs().add(1, 'hour').toDate(),
-    end: dayjs().add(2, 'hours').toDate(),
-    attendees: [
-      {
-        attendee: {
-          email: 'samyak3@rcal.com',
-        },
-      },
-    ],
-  };
 
-  const dummyevent2 = {
-    id: 1,
-    title: 'timed event',
-    start: dayjs().add(1, 'hour').toDate(),
-    end: dayjs().add(2, 'hours').toDate(),
-    attendees: [
-      {
-        attendee: {
-          email: 'samyak@rcal.com',
-        },
-      },
-      {
-        attendee: {
-          email: 'samyak2@rcal.com',
-        },
-      },
-      {
-        attendee: {
-          email: 'samyak3@rcal.com',
-        },
-      },
-    ],
-  };
 
-  const dummyevent3 = {
-    id: 1,
-    title: 'timed event',
-    start: dayjs().add(1, 'hour').toDate(),
-    end: dayjs().add(2, 'hours').toDate(),
-  };
+it('renders the event card', () => {
+  render(<EventCard title={'Marketing meet with John'} start={new Date()} end={new Date(new Date().setHours(new Date().getHours() + 2))} />);
 
-  const startDate = dayjs(dummyevent1.start);
-  const startTime = startDate.format('h:mm A');
-  const endTime = dayjs(dummyevent1.end).format('h:mm A');
+  const eventCardBox = screen.getByTestId('event-card-box');
+  const eventCardDates = screen.getByTestId('event-card-dates');
+  const eventCardContent = screen.getByTestId('event-card-content');
+  const eventCardTitle = screen.getByTestId('event-card-title');
 
-  it('should render the event name', () => {
-    const { getByText } = render(<EventCard {...dummyevent1} />);
-    const eventName = getByText(dummyevent1.title);
-    expect(eventName).toBeInTheDocument();
-  });
+  expect(eventCardBox).toBeInTheDocument();
+  expect(eventCardDates).toBeInTheDocument();
+  expect(eventCardContent).toBeInTheDocument();
+  expect(eventCardTitle).toBeInTheDocument();
+});
 
-  it('should render the number of attendees when there are less than or equal to one number of attendee', () => {
-    const { getByText } = render(<EventCard {...dummyevent1} />);
-    const attendeesText = `${dummyevent1.attendees.length} ${
-      dummyevent1.attendees.length > 1 ? 'Participants' : 'Participant'
-    }`;
-    const attendees = getByText(attendeesText);
-    expect(attendees).toBeInTheDocument();
-  });
+it('renders the event card attendees', () => {
+  render(
+    <EventCard
+      title={'Marketing meet with John'}
+      start={new Date()}
+      end={new Date(new Date().setHours(new Date().getHours() + 2))}
+      attendees={[{attendee: {email: 'test1@abc.com'}}, {attendee: {email: 'test2@abc.com'}}]}
+    />,
+  );
+  const eventCardattendees = screen.getByTestId('event-card-attendees');
+  expect(eventCardattendees).toBeInTheDocument();
+});
 
-  it('should render the number of attendees when there are more than one number of attendees', () => {
-    const { getByText } = render(<EventCard {...dummyevent2} />);
-    const attendeesText = `${dummyevent2.attendees.length} ${
-      dummyevent2.attendees.length > 1 ? 'Participants' : 'Participant'
-    }`;
-    const attendees = getByText(attendeesText);
-    expect(attendees).toBeInTheDocument();
-  });
+it('renders the event card location', () => {
+  render(
+    <EventCard
+      title={'Marketing meet with John'}
+      start={new Date()}
+      end={new Date(new Date().setHours(new Date().getHours() + 2))}
+      location={'Amsterdam'}
+    />,
+  );
+  const eventCardLocation = screen.getByTestId('event-card-location');
+  expect(eventCardLocation).toBeInTheDocument();
+});
 
-  it('should render the number of attendees when there are more than one number of attendees', () => {
-    const { getByText } = render(<EventCard {...dummyevent3} />);
-    const attendeesText = '0 Participant';
-    const attendees = getByText(attendeesText);
-    expect(attendees).toBeInTheDocument();
-  });
+it('renders the event card meet', () => {
+  render(
+    <EventCard
+      meetLink={'https://github.com/Real-Dev-Squad/website-calendar'}
+      title={'Marketing meet with John'}
+      start={new Date()}
+      end={new Date(new Date().setHours(new Date().getHours() + 2))}
+    />,
+  );
+  const eventCardMeet = screen.getByTestId('event-card-meet');
+  expect(eventCardMeet).toBeInTheDocument();
+});
 
-  it('should render the correct date', () => {
-    const { getByText } = render(<EventCard {...dummyevent1} />);
-    const dateText = date;
-    const eventDate = getByText(dateText);
-    expect(eventDate).toBeInTheDocument();
-  });
-
-  it('should render the correct month', () => {
-    const { getByText } = render(<EventCard {...dummyevent1} />);
-    const monthText = month;
-    const eventMonth = getByText(monthText);
-    expect(eventMonth).toBeInTheDocument();
-  });
-
-  it('should render the correct time', () => {
-    const { getByText } = render(<EventCard {...dummyevent1} />);
-    const timeText = `${startTime} — ${endTime}`;
-    const eventTime = getByText(timeText);
-    expect(eventTime).toBeInTheDocument();
-  });
+it('renders the event card no details', () => {
+  render(<EventCard title={'Marketing meet with John'} start={new Date()}
+  end={new Date(new Date().setHours(new Date().getHours() + 2))} />);
+  const eventCardNoDetails = screen.getByTestId('event-card-no-details');
+  expect(eventCardNoDetails).toBeInTheDocument();
 });
