+++ app/routes/events.tsx
@@ -2,7 +2,7 @@ import dayjs from 'dayjs';
 import { useState } from 'react';
 import Drawer from '~/components/common/drawer';
 import Navbar from '~/components/common/navbar';
-import EventCard from '~/components/events/eventCard';
+import EventCard from '~/components/common/eventCard';
 import { CalEvent } from '~/utils/interfaces';
 
 const Events = () => {
