+++ app/stories/socialEventCard.stories.tsx
@@ -1,16 +1,16 @@
 import { Story, Meta } from '@storybook/react';
+import SocialEventCard from '~/components/common/socialEventCard';
 import { CalEvent } from '~/utils/interfaces';
-import EventCard from '~/components/common/eventCard/index';
 
 export default {
-  title: 'Components/EventCard',
-  component: EventCard,
+  title: 'Components/SocialEvent',
+  component: SocialEventCard,
   argTypes: {
     title: { control: 'text' },
   },
 } as Meta;
 
-const Template: Story<CalEvent> = (args: CalEvent) => <EventCard {...args} />;
+const Template: Story<CalEvent> = (args: CalEvent) => <SocialEventCard {...args} />;
 
 export const Primary = Template.bind({});
 Primary.args = {
