+++ app/styles/tailwind.css
@@ -578,14 +578,6 @@ video {
   left: 0px;
 }
 
-.left-\[calc\(100\%\/3\)\] {
-  left: calc(100% / 3);
-}
-
-.top-\[calc\(87px\*2\)\] {
-  top: calc(87px * 2);
-}
-
 .right-0 {
   right: 0px;
 }
@@ -684,11 +676,6 @@ video {
 .my-6 {
   margin-top: 1.5rem;
   margin-bottom: 1.5rem;
-}
-
-.my-20 {
-  margin-top: 5rem;
-  margin-bottom: 5rem;
 }
 
 .mx-12 {
@@ -841,16 +828,12 @@ video {
   height: 1.5rem;
 }
 
-.h-\[calc\(87px\*3\)\] {
-  height: calc(87px * 3);
+.h-screen {
+  height: 100vh;
 }
 
 .h-full {
   height: 100%;
-}
-
-.h-screen {
-  height: 100vh;
 }
 
 .h-2 {
@@ -875,6 +858,10 @@ video {
 
 .h-16 {
   height: 4rem;
+}
+
+.h-4 {
+  height: 1rem;
 }
 
 .min-h-screen {
@@ -910,14 +897,6 @@ video {
   width: fit-content;
 }
 
-.w-\[calc\(100\%\/3\)\] {
-  width: calc(100% / 3);
-}
-
-.w-3 {
-  width: 0.75rem;
-}
-
 .w-80 {
   width: 20rem;
 }
@@ -948,6 +927,10 @@ video {
 
 .w-72 {
   width: 18rem;
+}
+
+.w-3 {
+  width: 0.75rem;
 }
 
 .min-w-fit {
@@ -1041,6 +1024,10 @@ video {
   grid-template-columns: repeat(1, minmax(0, 1fr));
 }
 
+.grid-cols-\[auto\2c 1fr\] {
+  grid-template-columns: auto 1fr;
+}
+
 .flex-row {
   flex-direction: row;
 }
@@ -1093,6 +1080,14 @@ video {
   gap: 1rem;
 }
 
+.gap-6 {
+  gap: 1.5rem;
+}
+
+.gap-1\.5 {
+  gap: 0.375rem;
+}
+
 .divide-y > :not([hidden]) ~ :not([hidden]) {
   --tw-divide-y-reverse: 0;
   border-top-width: calc(1px * calc(1 - var(--tw-divide-y-reverse)));
@@ -1141,12 +1136,12 @@ video {
   border-radius: 1rem;
 }
 
-.rounded-xl {
-  border-radius: 0.75rem;
-}
-
 .rounded {
   border-radius: 0.25rem;
+}
+
+.rounded-xl {
+  border-radius: 0.75rem;
 }
 
 .rounded-md {
@@ -1221,6 +1216,10 @@ video {
 
 .border-l-0 {
   border-left-width: 0px;
+}
+
+.border-l-2 {
+  border-left-width: 2px;
 }
 
 .border-solid {
@@ -1359,8 +1358,8 @@ video {
   --tw-bg-opacity: 0.4;
 }
 
-.p-10 {
-  padding: 2.5rem;
+.p-5 {
+  padding: 1.25rem;
 }
 
 .p-4 {
@@ -1414,11 +1413,6 @@ video {
   padding-bottom: 2.5rem;
 }
 
-.px-3 {
-  padding-left: 0.75rem;
-  padding-right: 0.75rem;
-}
-
 .px-12 {
   padding-left: 3rem;
   padding-right: 3rem;
@@ -1427,6 +1421,11 @@ video {
 .py-3 {
   padding-top: 0.75rem;
   padding-bottom: 0.75rem;
+}
+
+.px-3 {
+  padding-left: 0.75rem;
+  padding-right: 0.75rem;
 }
 
 .px-0 {
@@ -1477,6 +1476,14 @@ video {
 
 .pb-12 {
   padding-bottom: 3rem;
+}
+
+.pt-0\.5 {
+  padding-top: 0.125rem;
+}
+
+.pt-0 {
+  padding-top: 0px;
 }
 
 .text-left {
@@ -1561,12 +1568,12 @@ video {
   text-transform: capitalize;
 }
 
-.leading-5 {
-  line-height: 1.25rem;
-}
-
 .leading-10 {
   line-height: 2.5rem;
+}
+
+.leading-5 {
+  line-height: 1.25rem;
 }
 
 .leading-none {
